autotools: use autoconf for config.h and include it automatically
In this change config.h is made to be generated by autotools.  The header is
included using AM_CPPFLAGS, and that means it is included automatically to
all compilations before anything else is included.  That ensures various
configure dependant features work correctly.

Now when autotools work the standard way it is possible to simplify
great deal autogen.sh script.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>